@@ -16,7 +16,7 @@
 
 package org.springframework.cloud.stream.binder.rabbit.properties;
 
-import javax.validation.constraints.Min;
+import jakarta.validation.constraints.Min;
 
 import org.springframework.amqp.core.AcknowledgeMode;
 import org.springframework.amqp.core.MessageDeliveryMode;
